Move `AddColumnCommands` to `given` package
